pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["alex"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
mysqlclient = "2.1.0"
django-request-logging = "^0.7.5"
diploma-frontend = {path = "diploma-frontend/dist/diploma-frontend-0.6.tar.gz"}
colorlog = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
